requirements.txt

#to install this requirements.txt tap this command pip install -r requirements.txt 
asgiref==3.3.1
certifi==2020.12.5
Django>=3.1.5,<3.2.0
django-elasticsearch-dsl
elasticsearch==7.10.1
elasticsearch-dsl>=7.0.0,<8.0.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2


djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
PyJWT==2.3.0
rest-framework-simplejwt==0.0.2
